update dir middleware
